%get and %put syntax to transfer values
Syntax to allow transfer of variables between local and remote, rather
than their string representations.  This can be helpful if the local
side has resources such as plotting infrastructure which is not
available on the remote.

Get the value of expression `r = rhs` on the client and evaluate
`lhs = r` on the server:

    remote> %get lhs = rhs

Get the value of expression `r = rhs` on the server and evaluate
`lhs = r` on the client:

    remote> %put lhs = rhs